golang standard layout 적용
