+++ prism_back/.gitignore
@@ -1,1 +1,2 @@
 .env
+cmd.exe
