--- prism_back/Routers/OAuths/kakao/kakao.go
@@ -1,18 +1,0 @@
-package Kakao
-
-import (
-	"prism_back/Structs/Users/KakaoUser"
-
-	"github.com/gorilla/mux"
-)
-
-var kakaoUser = &KakaoUser.KakaoUser{}
-
-func RegisterHandlers(r *mux.Router) {
-    
-    r.HandleFunc("/login", kakaoUser.Login).Methods("GET")
-
-    r.HandleFunc("/with_token", kakaoUser.AfterProcess).Methods("GET")
-
-    r.HandleFunc("/logout", kakaoUser.Logout).Methods("GET")
-}
