--- prism_back/Routers/OAuths/oauth.go
@@ -1,16 +1,0 @@
-package OAuthRouter
-
-import (
-	"prism_back/Routers/OAuths/Kakao"
-	"prism_back/Structs/Users/KakaoUser"
-
-	"github.com/gorilla/mux"
-)
-
-var kakaoUser = &KakaoUser.KakaoUser{}
-
-func RegisterHandlers(r *mux.Router) {
-    
-	Kakao.RegisterHandlers(r.PathPrefix("/kakao").Subrouter())
-}
-
