--- prism_back/Structs/Users/KakaoUser/KakaoUser_test.go
@@ -1,1 +1,0 @@
-package KakaoUser
