+++ prism_back/internal/session/session.go
@@ -1,0 +1,15 @@
+package session
+
+import (
+	"os"
+
+	"github.com/gorilla/sessions"
+	"github.com/joho/godotenv"
+)
+
+var err = godotenv.Load(".env")
+var Store *sessions.CookieStore
+
+func SetupStore() {
+	Store = sessions.NewCookieStore([]byte(os.Getenv("SECRET_KEY")))
+}
