+++ prism_back/pkg/handlers/admin/admin.go
@@ -1,0 +1,24 @@
+package admin
+
+import (
+	"prism_back/pkg/handlers/admin/auth"
+	"prism_back/pkg/middleware/adminAccess"
+	Tech "prism_back/pkg/models/tech"
+
+	"github.com/gorilla/mux"
+)
+
+
+
+func RegisterHandlers(r *mux.Router) {
+	auth.RegisterHandlers(r.PathPrefix("/user").Subrouter())
+	adminRouter := r.PathPrefix("/access").Subrouter()
+	adminRouter.Use(adminAccess.AdminMiddleware)
+	AccessRequest(adminRouter)   
+}
+
+func AccessRequest(r *mux.Router) {
+    r.HandleFunc("/tech", Tech.Get_tech_list).Methods("GET")
+    r.HandleFunc("/tech", Tech.Post_tech_list).Methods("POST")
+    r.HandleFunc("/tech", Tech.Put_tech_list).Methods("PUT")
+}
