+++ prism_back/pkg/handlers/oauths/oauths.go
@@ -1,0 +1,14 @@
+package oauths
+
+import (
+	"prism_back/pkg/handlers/oauths/kakao"
+
+	"github.com/gorilla/mux"
+)
+
+
+func RegisterHandlers(r *mux.Router) {
+    
+	kakao.RegisterHandlers(r.PathPrefix("/kakao").Subrouter())
+}
+
