+++ prism_back/pkg/models/tech/tech_test.go
