+++ prism_back/pkg/models/user/kakaoUser/kakaoUser_test.go
@@ -1,0 +1,1 @@
+package kakaoUser
